Moved SearchFlightsRequest to domain package, deleted ErrorResponse class
