--- src/main/java/io/codelex/flightplanner/response/ErrorResponse.java
@@ -1,17 +1,0 @@
-package io.codelex.flightplanner.response;
-
-public class ErrorResponse {
-    private String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
